move entryGen script to `helpers` folder
